+++ server/src/routes.ts
@@ -1,13 +1,13 @@
 import express from 'express';
-import ProcessoController from '@controllers/ProcessoController';
+import ProcessController from '@controllers/ProcessController';
 
 const routes = express.Router();
-const processoController = new ProcessoController();
+const processController = new ProcessController();
 
-routes.post('/processo', processoController.create);
-routes.get('/processo', processoController.get);
-routes.delete('/processo/:id', processoController.delete);
-routes.put('/processo/:id', processoController.update);
+routes.post('/process', processController.create);
+routes.get('/process', processController.get);
+routes.delete('/process/:id', processController.delete);
+routes.put('/process/:id', processController.update);
 
 
 export default routes;
